+++ yandex/data_source_yandex_kubernetes_node_group.go
@@ -55,6 +55,20 @@ func dataSourceYandexKubernetesNodeGroup() *schema.Resource {
 				Computed: true,
 				Elem: &schema.Resource{
 					Schema: map[string]*schema.Schema{
+						"container_runtime": {
+							Type:     schema.TypeList,
+							Optional: true,
+							Computed: true,
+							MaxItems: 1,
+							Elem: &schema.Resource{
+								Schema: map[string]*schema.Schema{
+									"type": {
+										Type:     schema.TypeString,
+										Required: true,
+									},
+								},
+							},
+						},
 						"resources": {
 							Type:     schema.TypeList,
 							Computed: true,