+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 0.70.0 (Unreleased)
+FEATURES:
+* managed-kubernetes: add `container_runtime` attribute to `yandex_kubernetes_node_group` resource and data source
 
 ## 0.69.0 (December 27, 2021)
 FEATURES: